Extract edit quote form
